+++ CONTROLLER/Classes/Gerente.php
@@ -1,0 +1,64 @@
+<?php
+
+require("User.php");
+
+class Gerente extends User{
+    private $nomeCompleto;
+    private $dataNascimento;
+    private $email;
+    private $talefone;
+    private $foto;
+    //caminho da foto
+
+    function Gerente($cod_user){
+        $this->nomeCompleto = "";
+        $this->dataNascimento = "";
+        $this->email = "";
+        $this->telefone = "";
+        $this->foto = "";
+    }
+
+    function getNomeCompleto(){
+        return $this->nomeCompleto;
+    }
+
+    function getDataNascimento(){
+        return $this->dataNascimento;
+    }
+
+    function getEmail(){
+        return $this->email;
+    }
+
+    function getTelefone(){
+        return $this->telefone;
+    }
+
+    function getFoto(){
+        return $this->foto;
+        
+    }
+
+    function verCaixa(){
+        require("Caixa.php");
+    }
+
+    function verEstoque(){
+        require("Estoque.php");
+        //tela de estoque (Hermany)
+        header("Location: Estoque.php");
+    }
+
+    function CadastrarLoja(){
+        //dados da loja
+        require("Produto.php");
+    }
+
+    function RegistrarFuncionario(){
+        //cadastro 
+        require("Funcionario.php");
+    }
+}
+
+
+?>
