--- CONTROLLER/Logout.php
@@ -1,7 +1,0 @@
-<?php
-session_start();
-session_destroy();
-header("Location: TelaDeLogin.php");
-
-// destroi a sessão e redireciona para tela de login
-?>
