+++ CONTROLLER/Outros/Logout.php
@@ -1,0 +1,7 @@
+<?php
+session_start();
+session_destroy();
+header("Location: TelaDeLogin.php");
+
+// destroi a sessão e redireciona para tela de login
+?>
