+++ CONTROLLER/Transação/loginTrans.php
@@ -1,0 +1,3 @@
+<?php 
+
+?>
